requirements.txt
ollama==0.4.8
openai==1.76.0
python-dotenv==1.1.0
fastapi==0.115.12
langchain-ollama==0.3.2
streamlit==1.44.1
uvicorn
httpx